+++ app/views/devise/passwords/new.html.erb
@@ -13,4 +13,3 @@
   </div>
 <% end %>
 
-<%= render "devise/shared/links" %>
