+++ app/views/devise/sessions/new.html.erb
@@ -30,15 +30,15 @@
   <% end %>
 </div>
 
-<%= render "devise/shared/links" %>
+<hr>
+
+  <div class="external-login">
+    <%= render "devise/shared/links" %>
+  </div>
+  
 <div class="password_link">
 <h3>パスワードを変更</h3>
 
 <p><%= link_to "パスワード変更", new_user_password_path %></p>
 
-<p>
-<%- if devise_mapping.recoverable? && controller_name != 'passwords' && controller_name != 'registrations' %>
-  <%= link_to "Forgot your password?", new_password_path(resource_name) %><br />
-<% end %>
-</p>
 </div>
